Move FlowScanner method to search by StepDescriptor into test only to avoid a cyclic plugin dependency
